backend/requirements.txt
langgraph
langchain-openai
openai
tavily-python
dotenv
fastapi
uvicorn